+++ ElectronNET.API/Shell.cs
@@ -2,9 +2,8 @@
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using Newtonsoft.Json.Serialization;
-using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
+using ElectronNET.API.Extensions;
 
 namespace ElectronNET.API
 {
@@ -40,17 +39,14 @@ namespace ElectronNET.API
         /// <summary>
         /// Show the given file in a file manager. If possible, select the file.
         /// </summary>
-        /// <param name="fullPath"></param>
-        /// <returns>Whether the item was successfully shown.</returns>
-        public Task<bool> ShowItemInFolderAsync(string fullPath)
+        /// <param name="fullPath">The full path to the directory / file.</param>
+        public Task ShowItemInFolderAsync(string fullPath)
         {
-            var taskCompletionSource = new TaskCompletionSource<bool>();
+            var taskCompletionSource = new TaskCompletionSource<object>();
 
-            BridgeConnector.Socket.On("shell-showItemInFolderCompleted", (success) =>
+            BridgeConnector.Socket.On("shell-showItemInFolderCompleted", () =>
             {
                 BridgeConnector.Socket.Off("shell-showItemInFolderCompleted");
-
-                taskCompletionSource.SetResult((bool)success);
             });
 
             BridgeConnector.Socket.Emit("shell-showItemInFolder", fullPath);
@@ -59,22 +55,22 @@ namespace ElectronNET.API
         }
 
         /// <summary>
-        /// Open the given file in the desktop’s default manner.
+        /// Open the given file in the desktop's default manner.
         /// </summary>
-        /// <param name="fullPath"></param>
-        /// <returns>Whether the item was successfully opened.</returns>
-        public Task<bool> OpenItemAsync(string fullPath)
+        /// <param name="path">The path to the directory / file.</param>
+        /// <returns>The error message corresponding to the failure if a failure occurred, otherwise <see cref="string.Empty"/>.</returns>
+        public Task<string> OpenPathAsync(string path)
         {
-            var taskCompletionSource = new TaskCompletionSource<bool>();
+            var taskCompletionSource = new TaskCompletionSource<string>();
 
-            BridgeConnector.Socket.On("shell-openItemCompleted", (success) =>
+            BridgeConnector.Socket.On("shell-openPathCompleted", (errorMessage) =>
             {
-                BridgeConnector.Socket.Off("shell-openItemCompleted");
+                BridgeConnector.Socket.Off("shell-openPathCompleted");
 
-                taskCompletionSource.SetResult((bool)success);
+                taskCompletionSource.SetResult((string) errorMessage);
             });
 
-            BridgeConnector.Socket.Emit("shell-openItem", fullPath);
+            BridgeConnector.Socket.Emit("shell-openPath", path);
 
             return taskCompletionSource.Task;
         }
@@ -83,95 +79,50 @@ namespace ElectronNET.API
         /// Open the given external protocol URL in the desktop’s default manner. 
         /// (For example, mailto: URLs in the user’s default mail agent).
         /// </summary>
-        /// <param name="url"></param>
-        /// <returns>Whether an application was available to open the URL. 
-        /// If callback is specified, always returns true.</returns>
-        public Task<bool> OpenExternalAsync(string url)
+        /// <param name="url">Max 2081 characters on windows.</param>
+        /// <returns>The error message corresponding to the failure if a failure occurred, otherwise <see cref="string.Empty"/>.</returns>
+        public Task<string> OpenExternalAsync(string url)
         {
-            var taskCompletionSource = new TaskCompletionSource<bool>();
-
-            BridgeConnector.Socket.On("shell-openExternalCompleted", (success) =>
-            {
-                BridgeConnector.Socket.Off("shell-openExternalCompleted");
-
-                taskCompletionSource.SetResult((bool)success);
-            });
-
-            BridgeConnector.Socket.Emit("shell-openExternal", url);
-
-            return taskCompletionSource.Task;
+            return OpenExternalAsync(url, null);
         }
 
         /// <summary>
         /// Open the given external protocol URL in the desktop’s default manner. 
         /// (For example, mailto: URLs in the user’s default mail agent).
         /// </summary>
-        /// <param name="url"></param>
-        /// <param name="options">macOS only</param>
-        /// <returns>Whether an application was available to open the URL. 
-        /// If callback is specified, always returns true.</returns>
-        public Task<bool> OpenExternalAsync(string url, OpenExternalOptions options)
+        /// <param name="url">Max 2081 characters on windows.</param>
+        /// <param name="options">Controls the behavior of OpenExternal.</param>
+        /// <returns>The error message corresponding to the failure if a failure occurred, otherwise <see cref="string.Empty"/>.</returns>
+        public Task<string> OpenExternalAsync(string url, OpenExternalOptions options)
         {
-            var taskCompletionSource = new TaskCompletionSource<bool>();
+            var taskCompletionSource = new TaskCompletionSource<string>();
 
-            BridgeConnector.Socket.On("shell-openExternalCompleted", (success) =>
+            BridgeConnector.Socket.On("shell-openExternalCompleted", (error) =>
             {
                 BridgeConnector.Socket.Off("shell-openExternalCompleted");
 
-                taskCompletionSource.SetResult((bool)success);
+                taskCompletionSource.SetResult((string) error);
             });
 
-            BridgeConnector.Socket.Emit("shell-openExternal", url, JObject.FromObject(options, _jsonSerializer));
+            if (options == null)
+            {
+                BridgeConnector.Socket.Emit("shell-openExternal", url);
+            }
+            else
+            {
+                BridgeConnector.Socket.Emit("shell-openExternal", url, JObject.FromObject(options, _jsonSerializer));
+            }
 
             return taskCompletionSource.Task;
         }
 
         /// <summary>
-        /// Open the given external protocol URL in the desktop’s default manner. 
-        /// (For example, mailto: URLs in the user’s default mail agent).
+        /// Move the given file to trash and returns a <see cref="bool"/> status for the operation.
         /// </summary>
-        /// <param name="url"></param>
-        /// <param name="options">macOS only</param>
-        /// <param name="errorAction">Action to get the error message.</param>
-        /// <returns>Whether an application was available to open the URL. 
-        /// If callback is specified, always returns true.</returns>
-        public Task<bool> OpenExternalAsync(string url, OpenExternalOptions options, Action<Error> errorAction)
-        {
-            var taskCompletionSource = new TaskCompletionSource<bool>();
-
-            BridgeConnector.Socket.On("shell-openExternalCompleted", (success) =>
-            {
-                BridgeConnector.Socket.Off("shell-openExternalCompleted");
-
-                taskCompletionSource.SetResult((bool)success);
-            });
-
-            BridgeConnector.Socket.Off("shell-openExternalCallback");
-            BridgeConnector.Socket.On("shell-openExternalCallback", (args) => {
-                var urlKey = ((JArray)args).First.ToString();
-                var error = ((JArray)args).Last.ToObject<Error>();
-
-                if(_openExternalCallbacks.ContainsKey(urlKey))
-                {
-                    _openExternalCallbacks[urlKey](error);
-                }
-            });
-
-            _openExternalCallbacks.Add(url, errorAction);
-
-            BridgeConnector.Socket.Emit("shell-openExternal", url, JObject.FromObject(options, _jsonSerializer), true);
-
-            return taskCompletionSource.Task;
-        }
-
-        private Dictionary<string, Action<Error>> _openExternalCallbacks = new Dictionary<string, Action<Error>>();
-
-        /// <summary>
-        /// Move the given file to trash and returns a boolean status for the operation.
-        /// </summary>
-        /// <param name="fullPath"></param>
+        /// <param name="fullPath">The full path to the directory / file.</param>
+        /// <param name="deleteOnFail">Whether or not to unilaterally remove the item if the Trash is disabled or unsupported on the volume.</param>
         /// <returns> Whether the item was successfully moved to the trash.</returns>
-        public Task<bool> MoveItemToTrashAsync(string fullPath)
+        public Task<bool> MoveItemToTrashAsync(string fullPath, bool deleteOnFail)
         {
             var taskCompletionSource = new TaskCompletionSource<bool>();
 
@@ -179,10 +130,10 @@ namespace ElectronNET.API
             {
                 BridgeConnector.Socket.Off("shell-moveItemToTrashCompleted");
 
-                taskCompletionSource.SetResult((bool)success);
+                taskCompletionSource.SetResult((bool) success);
             });
 
-            BridgeConnector.Socket.Emit("shell-moveItemToTrash", fullPath);
+            BridgeConnector.Socket.Emit("shell-moveItemToTrash", fullPath, deleteOnFail);
 
             return taskCompletionSource.Task;
         }
@@ -198,9 +149,9 @@ namespace ElectronNET.API
         /// <summary>
         /// Creates or updates a shortcut link at shortcutPath.
         /// </summary>
-        /// <param name="shortcutPath"></param>
-        /// <param name="operation"></param>
-        /// <param name="options"></param>
+        /// <param name="shortcutPath">The path to the shortcut.</param>
+        /// <param name="operation">Default is <see cref="ShortcutLinkOperation.Create"/></param>
+        /// <param name="options">Structure of a shortcut.</param>
         /// <returns>Whether the shortcut was created successfully.</returns>
         public Task<bool> WriteShortcutLinkAsync(string shortcutPath, ShortcutLinkOperation operation, ShortcutDetails options)
         {
@@ -210,21 +161,20 @@ namespace ElectronNET.API
             {
                 BridgeConnector.Socket.Off("shell-writeShortcutLinkCompleted");
 
-                taskCompletionSource.SetResult((bool)success);
+                taskCompletionSource.SetResult((bool) success);
             });
 
-            BridgeConnector.Socket.Emit("shell-writeShortcutLink", shortcutPath, operation.ToString(), JObject.FromObject(options, _jsonSerializer));
+            BridgeConnector.Socket.Emit("shell-writeShortcutLink", shortcutPath, operation.GetDescription(), JObject.FromObject(options, _jsonSerializer));
 
             return taskCompletionSource.Task;
         }
 
         /// <summary>
         /// Resolves the shortcut link at shortcutPath.
-        /// 
         /// An exception will be thrown when any error happens.
         /// </summary>
-        /// <param name="shortcutPath"></param>
-        /// <returns></returns>
+        /// <param name="shortcutPath">The path tot the shortcut.</param>
+        /// <returns><see cref="ShortcutDetails"/> of the shortcut.</returns>
         public Task<ShortcutDetails> ReadShortcutLinkAsync(string shortcutPath)
         {
             var taskCompletionSource = new TaskCompletionSource<ShortcutDetails>();
@@ -233,7 +183,10 @@ namespace ElectronNET.API
             {
                 BridgeConnector.Socket.Off("shell-readShortcutLinkCompleted");
 
-                taskCompletionSource.SetResult((ShortcutDetails)shortcutDetails);
+                var shortcutObject = shortcutDetails as JObject;
+                var details = shortcutObject?.ToObject<ShortcutDetails>();
+
+                taskCompletionSource.SetResult(details);
             });
 
             BridgeConnector.Socket.Emit("shell-readShortcutLink", shortcutPath);
@@ -241,7 +194,7 @@ namespace ElectronNET.API
             return taskCompletionSource.Task;
         }
 
-        private JsonSerializer _jsonSerializer = new JsonSerializer()
+        private readonly JsonSerializer _jsonSerializer = new JsonSerializer()
         {
             ContractResolver = new CamelCasePropertyNamesContractResolver(),
             NullValueHandling = NullValueHandling.Ignore,