+++ ElectronNET.Host/api/app.ts
@@ -1,13 +1,20 @@
 let isQuitWindowAllClosed = true, electronSocket;
+let appWindowAllClosedEventId;
 export = (socket: SocketIO.Socket, app: Electron.App) => {
     electronSocket = socket;
-    // Quit when all windows are closed.
+
+    // By default, quit when all windows are closed
     app.on('window-all-closed', () => {
         // On macOS it is common for applications and their menu bar
         // to stay active until the user quits explicitly with Cmd + Q
-        if (process.platform !== 'darwin' &&
-            isQuitWindowAllClosed) {
+        if (process.platform !== 'darwin' && isQuitWindowAllClosed) {
             app.quit();
+        } else if (appWindowAllClosedEventId) {
+            // If the user is on macOS
+            // - OR -
+            // If the user has indicated NOT to quit when all windows are closed,
+            // emit the event.
+            electronSocket.emit('app-window-all-closed' + appWindowAllClosedEventId);
         }
     });
 
@@ -16,9 +23,7 @@ export = (socket: SocketIO.Socket, app: Electron.App) => {
     });
 
     socket.on('register-app-window-all-closed-event', (id) => {
-        app.on('window-all-closed', () => {
-            electronSocket.emit('app-window-all-closed' + id);
-        });
+        appWindowAllClosedEventId = id;
     });
 
     socket.on('register-app-before-quit-event', (id) => {
@@ -303,5 +308,13 @@ export = (socket: SocketIO.Socket, app: Electron.App) => {
 
     socket.on('appDockSetIcon', (image) => {
         app.dock.setIcon(image);
+    });
+
+    socket.on('appGetUserAgentFallback', () => {
+        electronSocket.emit('appGetUserAgentFallbackCompleted', app.userAgentFallback);
+    });
+
+    socket.on('appSetUserAgentFallback', (userAgent) => {
+        app.userAgentFallback = userAgent;
     });
 };