+++ ElectronNET.Host/api/shell.js
@@ -4,31 +4,29 @@ let electronSocket;
 module.exports = (socket) => {
     electronSocket = socket;
     socket.on('shell-showItemInFolder', (fullPath) => {
-        const success = electron_1.shell.showItemInFolder(fullPath);
-        electronSocket.emit('shell-showItemInFolderCompleted', success);
+        electron_1.shell.showItemInFolder(fullPath);
+        electronSocket.emit('shell-showItemInFolderCompleted');
     });
-    socket.on('shell-openItem', (fullPath) => {
-        const success = electron_1.shell.openItem(fullPath);
-        electronSocket.emit('shell-openItemCompleted', success);
+    socket.on('shell-openPath', async (path) => {
+        const errorMessage = await electron_1.shell.openPath(path);
+        electronSocket.emit('shell-openPathCompleted', errorMessage);
     });
-    socket.on('shell-openExternal', (url, options) => {
-        let success = true;
+    socket.on('shell-openExternal', async (url, options) => {
+        let result = "";
         if (options) {
-            electron_1.shell.openExternal(url, options).catch((error) => {
-                success = false;
-                electronSocket.emit('shell-openExternalCallback', [url, error]);
+            await electron_1.shell.openExternal(url, options).catch(e => {
+                result = e.message;
             });
         }
         else {
-            electron_1.shell.openExternal(url).catch((error) => {
-                success = false;
-                electronSocket.emit('shell-openExternalCallback', [url, error]);
+            await electron_1.shell.openExternal(url).catch((e) => {
+                result = e.message;
             });
         }
-        electronSocket.emit('shell-openExternalCompleted', success);
+        electronSocket.emit('shell-openExternalCompleted', result);
     });
-    socket.on('shell-moveItemToTrash', (fullPath) => {
-        const success = electron_1.shell.moveItemToTrash(fullPath);
+    socket.on('shell-moveItemToTrash', (fullPath, deleteOnFail) => {
+        const success = electron_1.shell.moveItemToTrash(fullPath, deleteOnFail);
         electronSocket.emit('shell-moveItemToTrashCompleted', success);
     });
     socket.on('shell-beep', () => {